Finish abstracting out trafficstats/ endpoint.
