--- routelandia/src/main/java/edu/pdx/its/portal/routelandia/JSONParser.java
@@ -1,60 +1,0 @@
-/*
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-*/
-
-package edu.pdx.its.portal.routelandia;
-
-import android.util.Log;
-import com.google.android.gms.maps.model.LatLng;
-
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.util.ArrayList;
-import java.util.List;
-import edu.pdx.its.portal.routelandia.entities.*;
-
-/**
- * Created by loc on 1/17/15.
- */
-public class JSONParser {
-    ArrayList<TravelingInfo> parseTravelingInfo(JSONObject jObj){
-        ArrayList<TravelingInfo> travelingInfoList = new ArrayList<>();
-
-        try {
-            JSONArray jResult = jObj.getJSONArray("results");
-            if (jResult == null) {
-                // No items were found? Sounds suspicious, but I guess we're done.
-                Log.i("RESULT", "Apparently nothing was in the results array...");
-                return travelingInfoList;
-            } else {
-                try {
-                    for (int i = 0; i < jResult.length(); i++) {
-                        //Create a JSONObject and use it to construct a TravelingInfo to add to the list
-                        JSONObject jsonObject = (JSONObject) jResult.get(i);
-                        travelingInfoList.add(new TravelingInfo(jsonObject));
-                    }
-                } catch (JSONException e) {
-                    e.printStackTrace();
-                }
-
-                return travelingInfoList;
-            }
-        } catch(JSONException je) {
-            // Going to log a message, but not abort the app... We'll just pretend there were no results.
-            Log.e("RESULT", "Could not parse result out of JSON array");
-            return travelingInfoList;
-        }
-    }
-}
