+++ routelandia/src/main/java/edu/pdx/its/portal/routelandia/entities/APIEntity.java
@@ -1,3 +1,17 @@
+/*
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
 package edu.pdx.its.portal.routelandia.entities;
 
 import android.util.Log;
@@ -142,4 +156,6 @@ public abstract class APIEntity {
     public <T> String getNestedEntityUrl(Class<T> klass) {
         return API_ROOT + getNestedEntityUrlComponent(klass);
     }
+
+
 }
