+++ routelandia/src/main/java/edu/pdx/its/portal/routelandia/entities/Highway.java
@@ -66,8 +66,4 @@ public class Highway extends APIEntity {
         return highwayid;
     }
 
-    public List<Station> getStations() {
-        return Station.fetchStationListForHighway(this.highwayid);
-    }
-
 }
