+++ routelandia/src/main/java/edu/pdx/its/portal/routelandia/entities/Station.java
@@ -146,64 +146,4 @@ public class Station extends APIEntity implements Serializable {
         latLngList.add(new LatLng(in.readDouble(), in.readDouble()));
     }
 
-
-
-    /**
-     * Get a list of all the stations available from the API.
-     *
-     * @return List<Station> All stations returned by the API
-     */
-    /*
-    public static List<Station> fetchStationList() {
-        List<Station> retVal = new ArrayList<>();
-        try{
-            JSONObject res = new ApiFetcher().execute(getListUrl()).get();
-            JSONArray stations = (JSONArray)res.get("results");
-            for (int i = 0; i <stations.length() ; i++) {
-                //Create a Station from the JSONObject for each array index
-                Station ts = new Station((JSONObject) stations.get(i));
-
-                //add the station to the list of stations
-                retVal.add(i, ts);
-            }
-        } catch (JSONException e) {
-            Log.e(TAG, e.getMessage());
-        } catch (InterruptedException e) {
-            Log.e(TAG, e.getMessage());
-        } catch (ExecutionException e) {
-            Log.e(TAG, e.getMessage());
-        }
-        return retVal;
-    }
-    */
-
-
-    /**
-     * Get a list of all the stations for a specific highway
-     *
-     * @param int highwayid
-     * @return List<Station> All the stations for the given highway
-     */
-    // TODO: This and fetchStationList method share a lot of code, and should be abstracted out to remove the duplication.
-    public static List<Station> fetchStationListForHighway(int highwayid) {
-        List<Station> retVal = new ArrayList<>();
-        try{
-            JSONObject res = ((APIResultWrapper)new ApiFetcher().execute(getListUrlNestedInHighway(highwayid)).get()).getParsedResponse();
-            JSONArray stations = (JSONArray)res.get("results");
-            for (int i = 0; i <stations.length() ; i++) {
-                //Create a Station from the JSONObject for each array index
-                Station ts = new Station((JSONObject) stations.get(i));
-
-                //add the station to the list of stations
-                retVal.add(i, ts);
-            }
-        } catch (JSONException e) {
-            Log.e(TAG, e.getMessage());
-        } catch (InterruptedException e) {
-            Log.e(TAG, e.getMessage());
-        } catch (ExecutionException e) {
-            Log.e(TAG, e.getMessage());
-        }
-        return retVal;
-    }
 }
